requirements.txt
torch~=2.0.1
numpy~=1.25.2
Pillow~=10.0.0
opencv-python~=4.8.0.74
facenet-pytorch~=2.5.2
torchvision~=0.15.2
pytest~=7.4.0
pymongo~=4.4.1